% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute.R
\name{deconvolute_cibersort}
\alias{deconvolute_cibersort}
\title{Deconvolute Mixture Data using CIBERSORT}
\usage{
deconvolute_cibersort(data, signature_matrix)
}
\arguments{
\item{data}{A data frame containing mixture expression data.}

\item{signature_matrix}{A data frame with the signature matrix.}
}
\value{
A tibble with deconvolution results.
}
\description{
Applies a custom CIBERSORT deconvolution on the input data using the provided signature matrix,
and transforms the output into a tibble.
}
