% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute.R
\name{deconvolute_cibersortx}
\alias{deconvolute_cibersortx}
\title{Deconvolute using CIBERSORTx}
\usage{
deconvolute_cibersortx(
  data,
  signature_matrix,
  perm = 1,
  rmbatch_S_mode = FALSE,
  source_GEPs = NULL,
  use_cibersortx = TRUE,
  rmbatch_B_mode = FALSE,
  QN = FALSE,
  absolute = FALSE,
  abs_method = "sig.score"
)
}
\arguments{
\item{data}{A data frame containing the mixture data to be deconvoluted.}

\item{signature_matrix}{A data frame containing the signature matrix.}

\item{perm}{An integer specifying the number of permutations to be performed. Default is 1.}

\item{rmbatch_S_mode}{A logical value indicating whether to remove batch effects in source GEPs mode. Default is FALSE.}

\item{source_GEPs}{A data frame containing the source gene expression profiles. Required if \code{rmbatch_S_mode} is TRUE.}

\item{use_cibersortx}{A logical value indicating whether to use CIBERSORTx. Default is TRUE.}

\item{rmbatch_B_mode}{A logical value indicating whether to remove batch effects in bulk mode. Default is FALSE.}

\item{QN}{A logical value indicating whether to perform quantile normalization. Default is FALSE.}

\item{absolute}{A logical value indicating whether to use absolute mode. Default is FALSE.}

\item{abs_method}{A character string specifying the method to use for absolute mode. Default is "sig.score".}
}
\value{
A data frame containing the deconvolution results.
}
\description{
This function performs deconvolution using the CIBERSORTx Docker image.
}
\details{
This function requires the CIBERSORTx Docker image to be installed. It also requires the user to set the \code{CIBERSORTX_EMAIL} and \code{CIBERSORTX_TOKEN} environment variables with their CIBERSORTx credentials.
}
