% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_symbols_map.R
\docType{data}
\name{gene_symbols_map}
\alias{gene_symbols_map}
\title{Gene Symbols Map}
\format{
\subsection{\code{gene_symbols_map}}{

A data frame with 103,478 rows and 2 columns:
\describe{
\item{Symbol}{Gene symbol}
\item{Approved.Symbol}{Approved gene symbol}
}
}
}
\source{
HGNC
}
\usage{
gene_symbols_map
}
\description{
Data of approved gene symbols.
}
\keyword{datasets}
