% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_gene_groups.R
\name{handle_gene_groups}
\alias{handle_gene_groups}
\title{Handle Gene Groups}
\usage{
handle_gene_groups(
  data,
  gene_column = "Genes",
  separator = ";",
  mode = "pick_first"
)
}
\arguments{
\item{data}{A data frame containing gene expression data. The specified gene column should contain gene groups separated by a specified separator.}

\item{gene_column}{A string specifying the name of the column containing gene groups. Default is "Genes".}

\item{separator}{A string specifying the separator used to split gene groups. Default is ";".}

\item{mode}{A string specifying the mode of handling gene groups. Currently, only "pick_first" is supported. Default is "pick_first".}
}
\value{
A data frame with the first gene in each group.
}
\description{
This function handles gene groups by picking the first gene in each group.
}
\details{
The function splits the gene groups in the specified column using the provided separator and picks the first gene in each group.
}
\examples{
\dontrun{
data <- data.frame(Genes = c("Gene1;Gene2", "Gene3;Gene4", "Gene5"), Sample1 = c(1, 2, 3))
result <- handle_gene_groups(data, gene_column = "Genes", separator = ";")
}

}
