% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_input_data.R
\name{handle_input_data}
\alias{handle_input_data}
\title{Handle Input Data}
\usage{
handle_input_data(x, gene_column = "Genes", as_tibble = TRUE)
}
\arguments{
\item{x}{An input data object. Supported types are SummarizedExperiment, tibble, or data frame.}

\item{gene_column}{A string specifying the name of the column containing gene identifiers. Default is "Genes".}

\item{as_tibble}{A logical value indicating whether to return a tibble with the first column being Genes (TRUE) or a data frame with the Genes column as rownames (FALSE). Default is TRUE.}
}
\value{
A data frame or tibble with the input data normalized to have a "Genes" column.
}
\description{
This function handles various input data types and normalizes them into a common format with a "Genes" column.
}
