% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_scaling.R
\name{handle_scaling}
\alias{handle_scaling}
\title{Handle Scaling of Protein Data}
\usage{
handle_scaling(data, gene_column = "Genes", unlog = TRUE, tpm = TRUE)
}
\arguments{
\item{data}{A data frame containing protein data. The first column should contain gene identifiers.}

\item{gene_column}{A string specifying the name of the column containing gene identifiers. Default is "Genes".}

\item{unlog}{A logical value indicating whether to unlog the data (convert from log2 scale). Default is TRUE.}

\item{tpm}{A logical value indicating whether to convert the data to TPM (Transcripts Per Million). Default is TRUE.}
}
\value{
A data frame with the scaled protein data.
}
\description{
This function handles the scaling of protein data by optionally unlogging and converting to TPM (Transcripts Per Million).
}
