% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoDeconv-package.R
\docType{package}
\name{proteoDeconv-package}
\alias{proteoDeconv}
\alias{proteoDeconv-package}
\title{proteoDeconv: Enabling Cell-Type Deconvolution of Proteomics Data}
\description{
Tools for deconvoluting proteomics data to identify and quantify cell types (e.g., immune cell types) in complex biological samples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ComputationalProteomics/proteoDeconv}
  \item Report bugs at \url{https://github.com/ComputationalProteomics/proteoDeconv/issues}
}

}
\author{
\strong{Maintainer}: Måns Zamore \email{mans.zamore@immun.lth.se} (\href{https://orcid.org/0000-0002-8108-1097}{ORCID})

}
\keyword{internal}
