% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute_bayesdebulk.R
\name{transform_bayesdebulk_output}
\alias{transform_bayesdebulk_output}
\title{Transform BayesDeBulk Output}
\usage{
transform_bayesdebulk_output(bayes_cell_fractions, ...)
}
\arguments{
\item{bayes_cell_fractions}{A matrix of cell fractions from BayesDeBulk.}

\item{...}{Additional arguments for \code{dplyr::mutate}.}
}
\value{
A tibble in long format with columns for sample, cell type, cell count, and method.
}
\description{
Transforms BayesDeBulk cell fraction output into a tidy long data frame.
}
\keyword{internal}
