% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_gene_symbols.R
\name{update_gene_symbols}
\alias{update_gene_symbols}
\title{Update Gene Symbols}
\usage{
update_gene_symbols(data, filepath = "", verbose = FALSE)
}
\arguments{
\item{data}{A data frame or vector containing gene symbols. If a data frame is provided, the first column is assumed to contain the gene symbols.}

\item{filepath}{A string specifying the file path for verbose output. Default is an empty string.}

\item{verbose}{A logical value indicating whether to print detailed messages. Default is FALSE.}
}
\value{
If \code{data} is a data frame, the function returns a data frame with updated gene symbols. If \code{data} is a vector, the function returns a vector of updated gene symbols.
}
\description{
This function updates gene symbols in a given dataset using a predefined gene symbols map.
}
\details{
The function uses the \code{HGNChelper} package to check and update gene symbols based on a predefined gene symbols map. If \code{verbose} is TRUE, the function prints detailed messages about the number of approved, not approved, and NA gene symbols.
}
\examples{
\dontrun{
data <- data.frame(Genes = c("BRCA1", "TP53", "EGFR"))
updated_data <- update_gene_symbols(data, verbose = TRUE)
}

}
