% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cibersortx_output.R
\name{convert_cibersortx_output}
\alias{convert_cibersortx_output}
\title{Convert CIBERSORTx Output}
\usage{
convert_cibersortx_output(data, file = NULL, filter = FALSE)
}
\arguments{
\item{data}{A data frame containing CIBERSORTx output.}

\item{file}{A character string specifying the file name (optional).}

\item{filter}{A logical value indicating whether to filter rows with P-value > 0.05.}
}
\value{
A data frame in long format.
}
\description{
This function converts the output from CIBERSORTx into a long format.
}
