% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_signature_matrix.R
\name{create_signature_matrix}
\alias{create_signature_matrix}
\title{Create Signature Matrix using CIBERSORTx}
\usage{
create_signature_matrix(
  refsample,
  phenoclasses,
  g_min = 50,
  g_max = 150,
  q_value = 0.3,
  replicates = 5,
  sampling = 0.5,
  fraction = 0.75,
  filter = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{refsample}{A data frame containing the reference profiles (with replicates). This parameter is required.}

\item{phenoclasses}{A data frame containing the cell type classes. This parameter is required.}

\item{g_min}{Minimum number of genes per cell type in the signature matrix. Default is 50.}

\item{g_max}{Maximum number of genes per cell type in the signature matrix. Default is 150.}

\item{q_value}{Q-value threshold for differential expression. Default is 0.3.}

\item{replicates}{Number of replicates to use for building the reference file. Default is 5.}

\item{sampling}{Fraction of available gene expression profiles selected using random sampling. Default is 0.5.}

\item{fraction}{Fraction of cells of the same identity showing evidence of expression. Default is 0.75.}

\item{filter}{Logical indicating whether to remove non-hematopoietic genes. Default is FALSE.}

\item{verbose}{Logical indicating whether to print detailed output. Default is TRUE.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A data frame containing the signature matrix.
}
\description{
This function generates a signature matrix from a reference sample and phenoclasses using the CIBERSORTx tool.
}
\details{
This function uses the CIBERSORTx Docker image to generate a signature matrix. It writes the input data to temporary files, constructs a Docker command, and executes it. The resulting signature matrix is read from the output directory and returned as a data frame.
}
\examples{
\dontrun{
refsample <- data.frame(...) # Your reference sample data
phenoclasses <- data.frame(...) # Your phenoclasses data
signature_matrix <- create_signature_matrix(refsample, phenoclasses)
}

}
