% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute_bayesdebulk.R
\name{get_signature_markers}
\alias{get_signature_markers}
\title{Get Signature Markers}
\usage{
get_signature_markers(Y, signature_matrix)
}
\arguments{
\item{Y}{A list of data matrices.}

\item{signature_matrix}{A matrix of signature marker values.}
}
\value{
A character matrix of markers.
}
\description{
Extracts unique signature markers from a signature matrix.
}
\keyword{internal}
