% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_gene_groups.R
\name{handle_gene_groups}
\alias{handle_gene_groups}
\title{Handle Protein/Gene Groups}
\usage{
handle_gene_groups(data, gene_column = "Genes", separator = ";")
}
\arguments{
\item{data}{A data frame containing protein data. The specified gene column should contain protein/gene groups separated by a specified separator.}

\item{gene_column}{A string specifying the name of the column containing gene groups. Default is "Genes".}

\item{separator}{A string specifying the separator used to split protein/gene groups. Default is ";".}
}
\value{
A data frame with the first protein/gene in each group.
}
\description{
This function handles protein/gene groups by picking the first protein/gene in each group.
}
\details{
The function splits the protein/gene groups in the specified column using the provided separator and picks the first protein/gene in each group.
}
